split out sitemap+summary.jsonld from datasets

+++ backend/app/routers/datasets.py
@@ -8,6 +8,7 @@ import zipfile
 from collections.abc import Mapping, Iterable
 from typing import List, Optional, Union
 
+import pymongo
 import pika
 from bson import ObjectId
 from bson import json_util
@@ -28,6 +29,11 @@ from rocrate.rocrate import ROCrate
 
 from app import dependencies
 from app import keycloak_auth
+from app.search.connect import (
+    connect_elasticsearch,
+    insert_record,
+    delete_document_by_id,
+)
 from app.config import settings
 from app.keycloak_auth import get_user, get_current_user
 from app.models.datasets import (
@@ -183,56 +189,31 @@ async def save_dataset(
     user=Depends(keycloak_auth.get_current_user),
     db: MongoClient = Depends(dependencies.get_db),
 ):
+    # Make connection to elatsicsearch
+    es = connect_elasticsearch()
+
+    # Check all connection and abort if any one of them is not available
+    if db is None or es is None:
+        raise HTTPException(status_code=503, detail="Service not available")
+        return
+
     result = dataset_in.dict()
     dataset_db = DatasetDB(**dataset_in.dict(), author=user)
     new_dataset = await db["datasets"].insert_one(dataset_db.to_mongo())
     found = await db["datasets"].find_one({"_id": new_dataset.inserted_id})
     dataset_out = DatasetOut.from_mongo(found)
+
+    # Add en entry to the dataset index
+    doc = {
+        "name": dataset_out.name,
+        "description": dataset_out.description,
+        "author": dataset_out.author.email,
+        "created": dataset_out.created,
+        "modified": dataset_out.modified,
+        "download": dataset_out.downloads,
+    }
+    insert_record(es, "dataset", doc, dataset_out.id)
     return dataset_out
-
-
-def is_str(v):
-    return type(v) is str
-
-
-schemaOrg_mapping = {
-    "id": "identifier",
-    "first_name": "givenName",
-    "last_name": "familyName",
-    "created": "dateCreated",
-    "modified": "dateModified",
-    "views": "interactionStatistic",
-    "downloads": "DataDownload",
-}
-
-
-def datasetout_str2jsonld(jstr):
-    "remap to schema.org key in a json str"
-    if not is_str(jstr):
-        jt = type(jstr)
-        print(f"str2jsonld:{jstr},wrong type:{jt}")
-        return None
-    global schemaOrg_mapping
-    for k, v in schemaOrg_mapping.items():
-        if k in jstr:
-            ks = f'"{k}":'
-            vs = f'"{v}":'
-            # print(f'replace:{ks} with:{vs}')
-            jstr = jstr.replace(ks, vs)
-    jstr = jstr.replace("{", '{"@context": {"@vocab": "https://schema.org/"},', 1)
-    return jstr
-
-
-def datasetout2jsonld(dso):
-    "dataset attributes as jsonld"
-    jstr = dso.json()
-    return datasetout_str2jsonld(jstr)
-
-
-def datasetout2jsonld_script(dso):
-    "dataset attributes in scrapable ld+json script"
-    jld = datasetout2jsonld(dso)
-    return f'<script type="application/ld+json">{jld}</script>'
 
 
 @router.get("", response_model=List[DatasetOut])
@@ -248,6 +229,7 @@ async def get_datasets(
         for doc in (
             await db["datasets"]
             .find({"author.email": user_id})
+            .sort([("created", pymongo.DESCENDING)])
             .skip(skip)
             .limit(limit)
             .to_list(length=limit)
@@ -255,7 +237,12 @@ async def get_datasets(
             datasets.append(DatasetOut.from_mongo(doc))
     else:
         for doc in (
-            await db["datasets"].find().skip(skip).limit(limit).to_list(length=limit)
+            await db["datasets"]
+            .find()
+            .sort([("created", pymongo.DESCENDING)])
+            .skip(skip)
+            .limit(limit)
+            .to_list(length=limit)
         ):
             datasets.append(DatasetOut.from_mongo(doc))
     return datasets
@@ -268,56 +255,6 @@ async def get_dataset(dataset_id: str, db: MongoClient = Depends(dependencies.ge
     ) is not None:
         return DatasetOut.from_mongo(dataset)
     raise HTTPException(status_code=404, detail=f"Dataset {dataset_id} not found")
-
-
-@router.get("/{dataset_id}/summary.jsonld", response_model=DatasetOut)
-async def get_dataset_jsonld(
-    dataset_id: str, db: MongoClient = Depends(dependencies.get_db)
-):
-    "get ld+json script for inside the dataset page, for scraping"
-    dso = get_dataset(dataset_id, db)
-    jlds = datasetout2jsonld_script(dso)
-    return jlds
-
-
-def put_txtfile(fn, s, wa="a"):
-    with open(fn, wa) as f:
-        return f.write(s)
-
-
-def datasets2sitemap(datasets):
-    "given an array of datasetObjs put out sitemap.xml"
-    top = """<?xml version="1.0" encoding="UTF-8"?>
-    <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-    """
-    sm = "sitemap.xml"  # could write to string and ret it all
-    if datasets and len(datasets) > 0:
-        put_txtfile(sm, top, "w")
-        URLb = settings.frontend_url
-        for ds in datasets:
-            objid = getattr(ds, "id")
-            if objid:
-                id = str(objid)
-                put_txtfile(sm, f"<url><loc>{URLb}/datasets/{id}</loc></url> ")
-        put_txtfile(sm, "</urlset>")
-
-
-# now the route
-
-
-def get_txtfile(fn):
-    "ret str from file"
-    with open(fn, "r") as f:
-        return f.read()
-
-
-@router.get("/sitemap.xml")
-async def sitemap() -> str:
-    datasets = get_datasets()
-    # could compare len(datasets) w/len of sitemap-file to see if could use cached one
-    datasets2sitemap(datasets)  # creates the sitemap.xml file, in case want to cache it
-    s = get_txtfile("sitemap.xml")
-    return s
 
 
 @router.get("/{dataset_id}/files")
@@ -417,7 +354,17 @@ async def delete_dataset(
     db: MongoClient = Depends(dependencies.get_db),
     fs: Minio = Depends(dependencies.get_fs),
 ):
+    # Make connection to elatsicsearch
+    es = connect_elasticsearch()
+
+    # Check all connection and abort if any one of them is not available
+    if db is None or fs is None or es is None:
+        raise HTTPException(status_code=503, detail="Service not available")
+        return
+
     if (await db["datasets"].find_one({"_id": ObjectId(dataset_id)})) is not None:
+        # delete from elasticsearch
+        delete_document_by_id(es, "dataset", dataset_id)
         # delete dataset first to minimize files/folder being uploaded to a delete dataset
 
         await db["datasets"].delete_one({"_id": ObjectId(dataset_id)})
